Refactor dataset handling in DatasetsContent and related components for improved functionality and code clarity
