--- DashAI/front/src/hooks/datasets/useNotebooksActions.js
@@ -1,36 +1,0 @@
-export function useNotebookActions({
-  fetchNotebooks,
-  selectNotebook,
-  clearSelectedNotebook,
-  clearSelectedDataset,
-  selectNotebookView,
-  selectDatasetView,
-  goToNotebookCreation,
-  createDatasetFromNotebook,
-}) {
-  const handleAddDatasetFromNotebook = async (name, selectedNotebook) => {
-    if (!selectedNotebook) return;
-
-    clearSelectedNotebook();
-    selectDatasetView();
-
-    createDatasetFromNotebook(name, selectedNotebook.id);
-  };
-
-  const handleNotebookCreated = async (createdNotebook) => {
-    await fetchNotebooks();
-    selectNotebookView();
-    selectNotebook(createdNotebook.id);
-    clearSelectedDataset();
-  };
-
-  const handleNewNotebookFromDataset = () => {
-    goToNotebookCreation();
-  };
-
-  return {
-    handleAddDatasetFromNotebook,
-    handleNotebookCreated,
-    handleNewNotebookFromDataset,
-  };
-}
